(Chatbox): Fix messages not marked as read
Prior to these changes, the user would have needed to click the message
field in order for the UI to be in focus. These changes use the Web
Visibility API to notify the backend when the WebView/UI is visible and
hence should mark the messages in that conversation as read.

The reason for using setTimeout is because of a debounce mechanism
implemented internally.

Signed-off-by: Victor Otieno <victor@slick.co.ke>