(Chatbox): Fix user activated link navigation.
When the user clicks on a link, the default behaviour should be that
the link in question is opened in the external browser.  Due to an
implementation detail in the inappwebview package, this wasn't the
case.

Specifying the launch mode in `launchUrl` was necessary as without
it, the link would still open in the webview.

Signed-off-by: Victor Otieno <victor@slick.co.ke>