(Notification.dart): Fix push notifications not working in background on Android
Fixes #42

This bug occured because the code previously assumed that the main isolate
would always be available even when the app has been terminated. When the
app has been terminated and a notification arrives, the handler is ran inside a
background isolate. No main isolate is started.

Hence the fix is to handle all the processing inside the background isolate.

While testing, it was noted that `_androidSettings` would be blank in the above
case. To fix this we encode the data and store in shared preferences.

Signed-off-by: Victor Otieno <victor@slick.co.ke>