(Github Action): Update publishing workflow
These changes are aimed at improving the securing the package publishing
process. The main change is the use of short-lived OIDC tokens signed by
Github instead of longlived tokens.

Also as part of this change is the enforcement of the use of commit SHAs
instead of version numbers for 3rd party actions used. This helps prevent
any supply chain attacks that may occur in future.

Signed-off-by: Victor Otieno <victor@slick.co.ke>